[Jinsujin/issue-tracker] Feat: NewIssueModel 계층 추가
